b6: ドメインの Conformist 化（委任 1）— 本家 v2.0.0 の trait を直接実装
ADR-010 の乗り換え第 1 段。event-store-adapter-rs = "=2.0.0"（完全固定）+ serde + chrono
（default-features = false）を導入し、ドメイン型が本家の trait を直接実装した。

- impl Event for WorkflowExecutionEvent — WorkflowExecutionEventId 新設
  （決定的採番: 集約 ID + seq_nr）、aggregate_id() / is_created()（genesis のみ true）/
  occurred_at() -> &DateTime<Utc> / seq_nr() -> usize
- impl Aggregate for WorkflowExecution — last_updated_at 新設（適用イベントの occurred_at
  で維持）/ seq_nr・version -> usize / serde derive（フィールドは private のまま）
- impl AggregateId for IntentId — type_name() / value()（外部 trait 準拠の例外、doc に理由）
- usize 化・chrono 化の波及（ワイヤ / ITF 準拠 / 契約テスト / FakeClock は Cell 化で
  expect 排除）。GlobalSeqNr は u64 据え置き（我々の横断カーソルのドメイン — 委任 2 で確定）
- with_version は削除（本家が version を採番するため）
- Published Language 面は逐語維持（ISO 8601 文字列面・ITF fixture・Quint 無変更、
  quint-gate 全緑で確認）

証明: 本家 memory バックエンドへの実 persist / 復元テスト 4 本（genesis 拒否の配線・
完全一致復元・楽観 version 採番・ジャーナルのみ追記から replay 追いつき）。
is_created を固定する mutation で 4 本中 3 本が落ちることを実測（検出力の証明）。

検査: 689 テスト全緑（674 + 15）/ fmt / clippy / cargo lint / quint-gate / cargo audit
両 lock 0 件。clippy 指摘 2 件は握り潰さず処理（BR1.7 射程外の 1 行 allow + 理由、
FakeClock は Cell 化で panic 経路ごと削除）。

規則の帰結 2 件を正本へ（Conformist 裁定からの導出）:
- upstream-contracts.md: 外部 trait の実装はこちらの命名規則の例外（要 doc 一行）。
  例外は trait が要求するメソッドだけ
- error-handling.md: thiserror/anyhow 不使用の射程は「我々が書くエラー型」。
  推移依存として入ることは違反ではない

設計質問 A〜C（Deserialize の不変条件迂回 / genesis version の意味論 / 横断読取の接続
所有者)は委任 2 の前にオーナー裁定。developer-report-1.md 参照。

+++ modules/core/domain/src/orchestration/apply_error.rs
@@ -10,9 +10,9 @@ pub enum ApplyError {
     /// 封筒の `seq_nr` が現在値 + 1 でない (BR2.1)。
     SequenceGap {
         /// 集約が期待した `seq_nr` (現在値 + 1)。
-        expected: u64,
+        expected: usize,
         /// イベント封筒が持っていた `seq_nr`。
-        actual: u64,
+        actual: usize,
     },
     /// ペイロードのステージ slug が `stages` に無い。
     UnknownStage(StageSlug),
